Add element grub2-boot-ipxe
Installs the package containing the ipxe.lkrn image and configures
grub to boot from it by default.

This will in theory allow an OVB environment to boot without needing a
patched nova in the host cloud. Modifying an existing grub based image
allows Ironic localboot to replace the contents of this image.

Change-Id: I7f3a380fdb5020b0caef851e2964762287a34558